UWP issue when no document is passed. (#114)
As a response to https://github.com/PSPDFKit/react-native/issues/112 an issue was found when not passing document from JS. I believe this should be an optional feature therefore this PR fixes this. 

No document will attempt to load by default if document is not present. 

In addition to this, some library alignment were made and copyrights updated.