Major Windows refactor (#126)
This is a PR to try and align windows with iOS and Android.

Callbacks and events called in a very similar way to Android, which is having a ref to the PdfView and calling through the UIManager. Not all functions from Android are implemented currently as the paradigms do not transfer.