Windows FTS  (#130)
The PR will add a new native module for the windows react native PSPDFKit layer.

LibraryModule exposes support for FTS in react native windows. You can open multiple libraries from either the picker or local folders in which the application has access to (use RNFS itinance/react-native-fs)

Configuration of the search is passed in via a json object which is converted internally and pass up to PSPDFKit windows. The results are passed back in a promise, the results comprises of a json objet with information dependent on the configuration. Much like the native library.