UWP Customize toolbar support (#165)
Fixes : https://github.com/PSPDFKit/react-native/issues/158

# Details
This PR brings Custom toolbar support to `react-native-pspdfkit` for UWP. `setToolbarItems` takes an array of `ToolbarItems` in which the format is described in same as in PSPDFKit for Web. Therefore the `ToolbarItem` is defined as https://pspdfkit.com/api/web/PSPDFKit.ToolbarItem.html
There are a set of supported default items listed https://pspdfkit.com/api/web/PSPDFKit.html#.defaultToolbarItems which can be referenced simply by setting the type. For example.
`this.refs.pdfView.setToolbarItems([{ type: "ink" }])` as seen in the Catalog example. 

It is also possible to retrieve the current toolbar items using `getToolbarItems` which will return with the same format as described above. 

There is also a fair amount of clean up in this PR as well as better support for 0.57 which had some deprecated methods.