Proposal for specific peer Dependencies (#178)
This adds specific peer Dependencies and therefore fixes #174

Also I would like a version update to mitigate #173 (Currently this PR does not change the version)

Are you fine with PRs which increment the patch version? The resolved package seems to be inconsistent if it's exactly the same version. (This seems to be a yarn or github issue, because the lock file should prevent this.)