Support css colors (#199)
Fixes #184

This will allow the users to set the 3 main colors used in the UI.

The props are now handled a little different, so we can do some custom processing. This way we can allow the users to pass in the extra colors as part of style and we split it out in index.windows.js.

In the future, it may be useful to allow the user to declare more css properties, or have an open css property to define any string, but validation on this will be hard and I imagine it's not a common use case.