Re-implement Css Customization (#221)
As explained in #219 (comment) due to limitations with file permissions in UWP it was necessary to rework the CSS customization. Therefore now the viable solution is to provide a native option.

In this PR I have reverted the previous CSS work and added the option to start the PSPDFKit react native package with a Uri parameter which will be passed as the CSS resource. Within this CSS resource, it's possible to set the main colors of the toolbar. As before.