Update to UWP 2.3.1 (#291)
It made sense to pull out the logic to check if a file is stage to be open into the view manager as we cannot guarantee that the PdfView is ready to go.
Along with this we do not track the open file in PdfViewPage but rely on PdfView to retrieve the document source. We throw and exception if anything other than a storage file is used because pspdfkit-react-native does not support this.

Removed an instance of showing a dialog, we now throw up to the promise.